Analytics_Bot/requirements.txt
mcp==1.22.0
fastmcp==2.13.1
langchain==1.0.8
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0

langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langgraph==1.0.3
chromadb==1.3.5
sentence-transformers==5.1.2
huggingface-hub==0.36.0
python-dotenv==1.1.1
sqlalchemy==2.0.44
psycopg2-binary==2.9.10
docx2txt==0.9
pandas==2.3.1
numpy==2.2.6
google-cloud-bigquery==3.37.0
uvicorn==0.38.0
Flask==3.1.1
requests==2.32.5
tiktoken==0.12.0
pydantic==2.11.7